implementaçao de paginaçao

+++ src/main/java/com/anjox/Gamebox_api/service/GameService.java
@@ -1,9 +1,15 @@
 package com.anjox.Gamebox_api.service;
 import com.anjox.Gamebox_api.dto.RequestCreateGameDto;
+import com.anjox.Gamebox_api.dto.ResponseGameDto;
+import com.anjox.Gamebox_api.dto.ResponsePaginationGameDto;
 import com.anjox.Gamebox_api.entity.GameEntity;
 import com.anjox.Gamebox_api.repository.GameRepository;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import java.util.stream.Collectors;
 
 
 @Service
@@ -30,23 +36,61 @@ public class GameService {
         return gameRepository.save(game);
     }
 
-    public GameEntity getGameById(Long id){
-        return gameRepository.findById(id).orElse(null);
+    public ResponseGameDto getGameById(Long id){
+        GameEntity game =  gameRepository.findById(id).orElse(null);
+        if(game!=null){
+            return new ResponseGameDto(
+                    game.getId(),
+                    game.getTitle(),
+                    game.getDescription(),
+                    game.getGenre(),
+                    game.getPrice(),
+                    game.getImageUrl()
+            ) ;
+        }
+        throw new RuntimeException("game nao encontrado");
     }
 
-    public List<GameEntity> getAllGames(){
-        return gameRepository.findAll();
+    public ResponsePaginationGameDto getAllGames(int size, int page){
+         Pageable pageable = PageRequest.of(page, size);
+         Page<GameEntity> games = gameRepository.findAll(pageable);
+        return getResponsePaginationGameDto(games);
     }
 
-    public List<GameEntity> getGamesByUserId(Long userId){
-        return gameRepository.findByUserId(userId);
+    public ResponsePaginationGameDto getGamesByUserId(Long userId, int page, int size){
+        Pageable pageable  = PageRequest.of(page, size);
+        Page<GameEntity> gamesByUserId=  gameRepository.findByUserId(userId, pageable);
+        return getResponsePaginationGameDto(gamesByUserId);
     }
 
     public void deleteGameById(Long id){
-        gameRepository.deleteById(id);
+        if(gameRepository.existsById(id)){
+            gameRepository.deleteById(id);
+        }
+        throw new RuntimeException("game nao encontrado");
     }
 
     public void deleteAllGamesByUserId(Long userId){
         gameRepository.deleteAllByUserId(userId);
+    }
+
+    private ResponsePaginationGameDto getResponsePaginationGameDto(Page<GameEntity> gamesByUserId) {
+        List<ResponseGameDto> games = gamesByUserId.getContent().stream().map(
+                g -> new ResponseGameDto(
+                        g.getId(),
+                        g.getTitle(),
+                        g.getDescription(),
+                        g.getGenre(),
+                        g.getPrice(),
+                        g.getImageUrl()
+                )
+        ).collect(Collectors.toList());
+        return new ResponsePaginationGameDto(
+                games,
+                gamesByUserId.getTotalPages(),
+                gamesByUserId.getTotalElements(),
+                gamesByUserId.getSize(),
+                gamesByUserId.getNumber()
+        );
     }
 }
